+++ packages/expo-modules-core/ios/JSI/EXJSIInstaller.h
@@ -20,4 +20,9 @@
  */
 + (BOOL)installExpoModulesHostObject:(nonnull EXAppContext *)appContext;
 
+/**
+ Installs the base class for shared objects, i.e. `global.expo.SharedObject`.
+ */
++ (void)installSharedObjectClass:(nonnull EXRuntime *)runtime releaser:(void(^)(long))releaser;
+
 @end